+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 - EditorLink.DetailEntity.detail() renamed get() for consistency.
 - AbstractEntityEditor.async() and ASYNC configuration value added, moved from SwingEntityEditor.
 - ModelLink.Builder.ModelStep added for consistency, related refactoring.
+- EditorLink.Builder R type parameter removed.
 ### is.codion.swing.framework.ui
 - EditorInspector now displays value and original value for attributes.
 - EntityEditPanel, input component selection now uses actual component label text instead of attribute caption, if available.